Added opentelemetry support and removed opentracing
