+++ modules/core/pom.xml
@@ -404,12 +404,23 @@
             <scope>test</scope>
         </dependency>
 
-        <!-- Jaeger Tracing -->
+		<!--OpenTelemetry-->
 		<dependency>
-			<groupId>org.wso2.orbit.io.jaeger</groupId>
-			<artifactId>jaeger-client</artifactId>
+			<groupId>org.wso2.orbit.io.opentelemetry</groupId>
+			<artifactId>opentelemetry-all</artifactId>
 		</dependency>
-		<!-- Zipkin Tracing -->
+		<dependency>
+			<groupId>org.wso2.orbit.com.squareup.okhttp</groupId>
+			<artifactId>okhttp</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.wso2.orbit.com.squareup.okio</groupId>
+			<artifactId>okio</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>io.zipkin.reporter2</groupId>
+			<artifactId>zipkin-sender-okhttp3</artifactId>
+		</dependency>
 		<dependency>
 			<groupId>io.zipkin.zipkin2</groupId>
 			<artifactId>zipkin</artifactId>
@@ -418,15 +429,6 @@
 			<groupId>io.zipkin.reporter2</groupId>
 			<artifactId>zipkin-reporter</artifactId>
 		</dependency>
-		<dependency>
-			<groupId>io.jaegertracing</groupId>
-			<artifactId>jaeger-zipkin</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>io.zipkin.reporter2</groupId>
-			<artifactId>zipkin-sender-urlconnection</artifactId>
-		</dependency>
-
 
         <dependency>
             <groupId>com.github.stefanbirkner</groupId>
