+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/collectors/CloseEventCollector.java
@@ -26,7 +26,7 @@ import org.apache.synapse.aspects.flow.statistics.data.raw.BasicStatisticDataUni
 import org.apache.synapse.aspects.flow.statistics.data.raw.StatisticDataUnit;
 import org.apache.synapse.aspects.flow.statistics.log.templates.EndFlowEvent;
 import org.apache.synapse.aspects.flow.statistics.log.templates.StatisticsCloseEvent;
-import org.apache.synapse.aspects.flow.statistics.opentracing.OpenTracingManagerHolder;
+import org.apache.synapse.aspects.flow.statistics.tracing.opentelemetry.OpenTelemetryManagerHolder;
 import org.apache.synapse.aspects.flow.statistics.util.StatisticDataCollectionHelper;
 import org.apache.synapse.aspects.flow.statistics.util.StatisticsConstants;
 
@@ -95,8 +95,8 @@ public class CloseEventCollector extends RuntimeStatisticCollector {
 				addEventAndDecrementCount(messageContext, closeEvent);
 			}
 
-			if (isOpenTracingEnabled()) {
-				OpenTracingManagerHolder.getOpenTracingManager().getHandler().
+			if (isOpenTelemetryEnabled()) {
+				OpenTelemetryManagerHolder.getOpenTelemetryManager().getHandler().
 						handleCloseEntryEvent(statisticDataUnit, messageContext);
 			}
 
@@ -124,9 +124,9 @@ public class CloseEventCollector extends RuntimeStatisticCollector {
                 addEventAndCloseFlow(messageContext, endFlowEvent);
             }
 
-            if (isOpenTracingEnabled()) {
-				OpenTracingManagerHolder.getOpenTracingManager().getHandler()
-					.handleCloseFlowForcefully(dataUnit, messageContext);
+            if (isOpenTelemetryEnabled()) {
+				OpenTelemetryManagerHolder.getOpenTelemetryManager().getHandler()
+						.handleCloseFlowForcefully(dataUnit, messageContext);
 			}
 
         }
