--- modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/opentracing/management/OpenTracingManager.java
@@ -1,39 +1,0 @@
-/*
- *  Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package org.apache.synapse.aspects.flow.statistics.opentracing.management;
-
-import org.apache.synapse.aspects.flow.statistics.opentracing.management.handling.span.OpenTracingSpanHandler;
-
-/**
- * The interface for the manager, which coordinates OpenTracing.
- */
-public interface OpenTracingManager {
-
-    /**
-     * Resolves the OpenTracing compatible span handler.
-     */
-    void resolveHandler();
-
-    /**
-     * Returns the OpenTracing compatible span handler.
-     * @return An OpenTracing compatible span handler.
-     */
-    OpenTracingSpanHandler getHandler();
-
-}
