--- modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/opentracing/management/handling/event/ContinuationStateStackEventHandler.java
@@ -1,51 +1,0 @@
-/*
- *  Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package org.apache.synapse.aspects.flow.statistics.opentracing.management.handling.event;
-
-import org.apache.synapse.ContinuationState;
-import org.apache.synapse.MessageContext;
-import org.apache.synapse.SequenceType;
-
-/**
- * The interface for handling modifications related to the continuation state stack,
- * that were done by the ContinuationStateStackManager.
- */
-public interface ContinuationStateStackEventHandler {
-
-    /**
-     * Handles insertion to the continuation state stack.
-     * @param synCtx    Message context.
-     * @param seqName   Sequence name.
-     * @param seqType   Sequence type.
-     */
-    void handleStateStackInsertion(MessageContext synCtx, String seqName, SequenceType seqType);
-
-    /**
-     * Handles removal from the continuation state stack.
-     * @param continuationState Continuation state.
-     * @param synCtx            Message context.
-     */
-    void handleStateStackRemoval(ContinuationState continuationState, MessageContext synCtx);
-
-    /**
-     * Handles clearance of the continuation state stack.
-     * @param synCtx    Message context.
-     */
-    void handleStateStackClearance(MessageContext synCtx);
-}
