+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/tracing/opentelemetry/management/handling/span/OpenTelemetrySpanHandler.java
@@ -1,0 +1,32 @@
+/*
+ *  Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+
+package org.apache.synapse.aspects.flow.statistics.tracing.opentelemetry.management.handling.span;
+
+import org.apache.synapse.aspects.flow.statistics.tracing.opentelemetry.management.handling.event.CallbackEventHandler;
+import org.apache.synapse.aspects.flow.statistics.tracing.opentelemetry.management.handling.event.CloseEventHandler;
+import org.apache.synapse.aspects.flow.statistics.tracing.opentelemetry.management.handling.event.ContinuationStateStackEventHandler;
+import org.apache.synapse.aspects.flow.statistics.tracing.opentelemetry.management.handling.event.OpenEventHandler;
+
+/**
+ * Interface for the controller which handles OpenTracing compatible spans, with respect to various events received
+ * during Synapse message flow.
+ */
+public interface OpenTelemetrySpanHandler
+        extends OpenEventHandler, CloseEventHandler, CallbackEventHandler, ContinuationStateStackEventHandler {
+}
