+++ pom.xml
@@ -1104,16 +1104,26 @@
               <artifactId>libphonenumber</artifactId>
               <version>${com.googlecode.libphonenumber.version}</version>
          </dependency>
-          <!--Jaeger Tracing-->
+          <!--OpenTelemetry-->
           <dependency>
-              <groupId>org.wso2.orbit.io.jaeger</groupId>
-              <artifactId>jaeger-client</artifactId>
-              <version>${jaeger-client.version}</version>
+              <groupId>org.wso2.orbit.io.opentelemetry</groupId>
+              <artifactId>opentelemetry-all</artifactId>
+              <version>${opentelemetry.all.version}</version>
+          </dependency>
+          <dependency>
+              <groupId>org.wso2.orbit.com.squareup.okhttp</groupId>
+              <artifactId>okhttp</artifactId>
+              <version>${okhttp.wso2.version}</version>
+          </dependency>
+          <dependency>
+              <groupId>org.wso2.orbit.com.squareup.okio</groupId>
+              <artifactId>okio</artifactId>
+              <version>${okio.wso2.version}</version>
           </dependency>
           <dependency>
               <groupId>io.zipkin.zipkin2</groupId>
               <artifactId>zipkin</artifactId>
-              <version>${zipkin-client.version}</version>
+              <version>${zipkin.version}</version>
           </dependency>
           <dependency>
               <groupId>io.zipkin.reporter2</groupId>
@@ -1122,13 +1132,8 @@
           </dependency>
           <dependency>
               <groupId>io.zipkin.reporter2</groupId>
-              <artifactId>zipkin-sender-urlconnection</artifactId>
-              <version>${zipkin-urlconnection.version}</version>
-          </dependency>
-          <dependency>
-              <groupId>io.jaegertracing</groupId>
-              <artifactId>jaeger-zipkin</artifactId>
-              <version>${jaeger-zipkin.version}</version>
+              <artifactId>zipkin-sender-okhttp3</artifactId>
+              <version>${zipkin.sender.okhttp3.version}</version>
           </dependency>
           <!--Mockito Dependency-->
           <dependency>
@@ -1547,11 +1552,12 @@
        <transport.http.netty.version>6.3.35</transport.http.netty.version>
        <version.org.wso2.orbit.javax.activation>1.1.1.wso2v3</version.org.wso2.orbit.javax.activation>
        <rabbitmq.version>5.8.0</rabbitmq.version>
-       <jaeger-client.version>1.8.0.wso2v1</jaeger-client.version>
-       <zipkin-client.version>2.23.1</zipkin-client.version>
-       <zipkin-reporter.version>2.16.2</zipkin-reporter.version>
-       <zipkin-urlconnection.version>2.16.2</zipkin-urlconnection.version>
-       <jaeger-zipkin.version>1.5.0</jaeger-zipkin.version>
+       <opentelemetry.all.version>1.11.0.wso2v2</opentelemetry.all.version>
+       <okhttp.wso2.version>4.9.3.wso2v1</okhttp.wso2.version>
+       <okio.wso2.version>2.8.0.wso2v1</okio.wso2.version>
+       <zipkin.sender.okhttp3.version>2.16.3</zipkin.sender.okhttp3.version>
+       <zipkin.version>2.23.2</zipkin.version>
+       <zipkin-reporter.version>2.16.3</zipkin-reporter.version>
        <javax.annotation.verion>1.3.2</javax.annotation.verion>
        <apache.felix.scr.ds-annotations.version>1.2.4</apache.felix.scr.ds-annotations.version>
        <com.hierynomus.smbj.version>0.11.3.wso2v1</com.hierynomus.smbj.version>
